Add files via upload
Completed: A file in which I've created all the tables for our schema and included the columns names and parameters when creating the tables using python. Code is running with no errors. 
Notr: Need the 'create_db.py' file and 'db.sqlite3' file in same directory to run the script. 

Working On: Logic behind creation and deletion of specific tables when reconnecting database. 
Ex) The 'Users' table should not be dropped, since once a user creates an account, it is permanently there, until a user desires deletion of their account information. 
      However, for the 'Picture' table, table should be dropped after closing connection, and then new 'Picture' table can be created when reconnecting database, where 
      the picture data is extracted from a csv file once again. This would help avoid duplicate picture listings in the 'Picture' table otherwise.