Update create_tables_with_header.py
Completed: Changed some of the table attribute names and constraints of values.

Again, this is a file in which I've created all the tables for our schema and included the columns names and parameters when creating the tables using python. 
Note: Need the 'create_db.py' file and 'db.sqlite3' file in same directory to run the script. 



Still Working On: Logic behind creation and deletion of specific tables when reconnecting database. 
------------------------------------------------------------------------------------------------------
Ex) The 'Users' table should not be dropped, since once a user creates an account, it is permanently there, until a user desires deletion of their account information.
Ex) However, for the 'Picture' table, table should be dropped after closing connection, and then new 'Picture' table can be created when reconnecting database, where the picture data is extracted from a csv file once again. This would help avoid duplicate picture listings in the 'Picture' table